infrastructure/aws/cdk/requirements.txt
aws-cdk-lib==2.80.0
constructs>=10.0.0,<11.0.0
aws-cdk.aws-apigateway>=2.80.0
aws-cdk.aws-cognito>=2.80.0
aws-cdk.aws-lambda>=2.80.0
aws-cdk.aws-iam>=2.80.0
aws-cdk.aws-ec2>=2.80.0
aws-cdk.aws-ecs>=2.80.0
aws-cdk.aws-rds>=2.80.0
pytest>=6.0.0
